Refactor Staged File Diff View
Renamed `StageFileDiffView` to `StageFileDiffsView` to reflect the handling of multiple file diffs. Removed the old `StageFileDiffView.swift` file and created a new `StageFileDiffsView.swift` file with updated implementation. Modified related references in `project.pbxproj` and `StagedView.swift`.
